+++ src/crypto/asn1/tasn_utl.c
@@ -200,7 +200,7 @@ int asn1_enc_restore(int *len, unsigned char **out, ASN1_VALUE **pval,
     *out += enc->len;
   }
   if (len) {
-    *len = enc->len;
+    *len = (int)(enc->len);
   }
   return 1;
 }
